# backend/src/main/java/com/diettracker/admin/dto/AdminLoginRequest.java
package com.diettracker.admin.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record AdminLoginRequest(
        @NotBlank @Size(min = 3, max = 64) String username,
        @NotBlank @Size(min = 8, max = 128) String password) {
}