# docs/admin/openapi.yaml
openapi: 3.1.0
info:
  title: 轻养管理 API
  version: 0.1.0-a0
  description: A0 冻结、A1/A2 实现范围。所有管理 Token 的 audience 为 diet-tracker-admin。
servers:
  - url: /api/admin
security:
  - adminBearer: []
paths:
  /auth/login:
    post:
      security: []
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminLoginRequest'}
      responses:
        '200': {description: 登录成功, content: {application/json: {schema: {$ref: '#/components/schemas/AdminSession'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /auth/me:
    get:
      operationId: getAdminSession
      responses:
        '200': {description: 当前管理员, content: {application/json: {schema: {$ref: '#/components/schemas/AdminProfile'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auth/logout:
    post:
      operationId: adminLogout
      responses:
        '204': {description: 会话已失效}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /food-categories:
    get:
      operationId: listAdminFoodCategories
      responses:
        '200': {description: 分类列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/FoodCategory'}}}}}
    post:
      operationId: createAdminFoodCategory
      parameters: [{$ref: '#/components/parameters/AuditReason'}]
      requestBody: {$ref: '#/components/requestBodies/FoodCategoryInput'}
      responses:
        '201': {description: 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/FoodCategory'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
  /food-categories/{id}:
    put:
      operationId: updateAdminFoodCategory
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/AuditReason'}]
      requestBody: {$ref: '#/components/requestBodies/FoodCategoryInput'}
      responses:
        '200': {description: 已更新, content: {application/json: {schema: {$ref: '#/components/schemas/FoodCategory'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      operationId: deleteAdminFoodCategory
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/AuditReason'}]
      responses:
        '204': {description: 已删除}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {description: 分类仍被系统食品引用, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
  /foods:
    get:
      operationId: listAdminFoods
      parameters:
        - {name: keyword, in: query, schema: {type: string, maxLength: 50}}
        - {name: categoryId, in: query, schema: {type: integer, format: int64}}
        - {name: source, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
        - {name: sort, in: query, schema: {type: string, default: updatedAt, enum: [name, updatedAt, calories]}}
        - {name: direction, in: query, schema: {type: string, default: desc, enum: [asc, desc]}}
      responses:
        '200': {description: 系统食品分页, content: {application/json: {schema: {$ref: '#/components/schemas/FoodPage'}}}}
    post:
      operationId: createAdminFood
      parameters: [{$ref: '#/components/parameters/AuditReason'}]
      requestBody: {$ref: '#/components/requestBodies/FoodInput'}
      responses:
        '201': {description: 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/Food'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
  /foods/{id}:
    put:
      operationId: updateAdminFood
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/AuditReason'}]
      requestBody: {$ref: '#/components/requestBodies/FoodInput'}
      responses:
        '200': {description: 已更新, content: {application/json: {schema: {$ref: '#/components/schemas/Food'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /foods/preview:
    post:
      operationId: previewAdminFoodNutrition
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FoodPreviewRequest'}
      responses:
        '200': {description: 同口径预览, content: {application/json: {schema: {$ref: '#/components/schemas/NutritionPreview'}}}}
  /custom-foods:
    get:
      operationId: diagnoseCustomFoods
      parameters:
        - {name: userRef, in: query, required: true, schema: {type: string}}
        - {name: foodId, in: query, required: true, schema: {type: integer, format: int64}}
      responses:
        '200': {description: 用户私有食品只读诊断, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/CustomFoodDiagnostic'}}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
components:
  securitySchemes:
    adminBearer: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: integer, format: int64}}
    AuditReason: {name: X-Audit-Reason, in: header, required: true, description: UTF-8 文本须使用百分号编码，服务端解码后按 2-200 字校验并审计, schema: {type: string}}
  requestBodies:
    FoodCategoryInput:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/FoodCategoryInput'}}}
    FoodInput:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/FoodInput'}}}
  responses:
    Unauthorized: {description: 未登录或会话失效, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Forbidden: {description: 权限不足, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    NotFound: {description: 资源不存在, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    RateLimited: {description: 登录尝试过于频繁, content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
  schemas:
    AdminLoginRequest: {type: object, required: [username, password], properties: {username: {type: string, minLength: 3, maxLength: 64}, password: {type: string, minLength: 8, maxLength: 128}}}
    AdminProfile: {type: object, required: [id, username, displayName, role], properties: {id: {type: integer, format: int64}, username: {type: string}, displayName: {type: string}, role: {$ref: '#/components/schemas/AdminRole'}}}
    AdminSession: {allOf: [{$ref: '#/components/schemas/AdminProfile'}, {type: object, required: [token, expiresIn], properties: {token: {type: string}, expiresIn: {type: integer, format: int64}}}]}
    AdminRole: {type: string, enum: [SUPER_ADMIN, FOOD_EDITOR, SUPPORT_VIEWER, RULE_PUBLISHER, ANALYST]}
    FoodCategoryInput: {type: object, required: [name, sortOrder], properties: {name: {type: string, minLength: 1, maxLength: 50}, icon: {type: string, maxLength: 100}, sortOrder: {type: integer, minimum: 0, maximum: 9999}}}
    FoodCategory: {allOf: [{$ref: '#/components/schemas/FoodCategoryInput'}, {type: object, required: [id, systemFoodCount, updatedAt], properties: {id: {type: integer, format: int64}, systemFoodCount: {type: integer}, updatedAt: {type: string, format: date-time}}}]}
    FoodInput:
      type: object
      required: [name, categoryId, baseAmount, baseUnit, calories, protein, fat, carbs]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        categoryId: {type: integer, format: int64}
        baseAmount: {type: number, exclusiveMinimum: 0}
        baseUnit: {type: string, minLength: 1, maxLength: 20}
        servingAmount: {type: [number, 'null'], exclusiveMinimum: 0}
        servingUnit: {type: [string, 'null'], maxLength: 20}
        calories: {type: number, minimum: 0}
        protein: {type: number, minimum: 0}
        fat: {type: number, minimum: 0}
        carbs: {type: number, minimum: 0}
        source: {type: string, enum: [SYSTEM, SYSTEM_SEED, SYSTEM_EDITORIAL]}
    Food: {allOf: [{$ref: '#/components/schemas/FoodInput'}, {type: object, required: [id, updatedAt], properties: {id: {type: integer, format: int64}, categoryName: {type: string}, updatedAt: {type: string, format: date-time}, lastOperator: {type: [string, 'null']}}}]}
    FoodPage: {type: object, required: [items, page, size, total, hasNext], properties: {items: {type: array, items: {$ref: '#/components/schemas/Food'}}, page: {type: integer}, size: {type: integer}, total: {type: integer, format: int64}, hasNext: {type: boolean}}}
    FoodPreviewRequest: {type: object, required: [food, amount], properties: {food: {$ref: '#/components/schemas/FoodInput'}, amount: {type: number, minimum: 1, maximum: 10000, default: 150}}}
    NutritionPreview: {type: object, required: [amount, calories, protein, fat, carbs], properties: {amount: {type: number}, calories: {type: number}, protein: {type: number}, fat: {type: number}, carbs: {type: number}}}
    CustomFoodDiagnostic: {type: object, required: [id, userRef, name, baseAmount, baseUnit, calories, protein, fat, carbs], properties: {id: {type: integer, format: int64}, userRef: {type: string, description: 不可逆脱敏摘要}, name: {type: string}, baseAmount: {type: number}, baseUnit: {type: string}, calories: {type: number}, protein: {type: number}, fat: {type: number}, carbs: {type: number}, updatedAt: {type: string, format: date-time}}}
    ApiError: {type: object, required: [code, message, requestId, fieldErrors], properties: {code: {type: string}, message: {type: string}, requestId: {type: string}, fieldErrors: {type: object, additionalProperties: {type: string}}}}